[IMP] hr_timesheet: correctly override graph model in task
This commit reviews the custom graph views in project/timesheets app to
make sure the graph of task no longer uses the custom graph view of
timesheets for instance. By doing separate custom views, it would be
easier to adapt a specific graph view.

task-4801249
