+++ addons/hr_timesheet/report/project_report_view.xml
@@ -6,9 +6,6 @@
             <field name="model">report.project.task.user</field>
             <field name="inherit_id" ref="project.view_task_project_user_graph" />
             <field name="arch" type="xml">
-                <xpath expr="//graph" position="attributes">
-                    <attribute name="js_class">hr_timesheet_project_task_analysis_graph</attribute>
-                </xpath>
                 <xpath expr="//field[@name='project_id']" position='after'>
                     <field name="allocated_hours" widget="timesheet_uom" type="measure"/>
                     <field name="effective_hours" widget="timesheet_uom" type="measure"/>
