+++ addons/hr_timesheet/static/src/views/project_task_analysis_graph/project_task_analysis_graph_model.js
@@ -1,0 +1,5 @@
+import { ProjectTaskAnalysisGraphModel } from "addons/project/static/src/views/project_task_analysis_graph/project_task_analysis_graph_model";
+import { patchGraphModel } from "../graph_model_patch";
+
+patchGraphModel(ProjectTaskAnalysisGraphModel);
+
