--- addons/hr_timesheet/static/src/views/project_task_analysis_graph/project_task_analysis_graph_view.js
@@ -1,8 +1,0 @@
-import { registry } from "@web/core/registry";
-import { hrTimesheetGraphView } from "../timesheet_graph/timesheet_graph_view";
-import { ProjectTaskAnalysisGraphRenderer } from "@project/views/project_task_analysis_graph/project_task_analysis_graph_renderer";
-
-registry.category("views").add("hr_timesheet_project_task_analysis_graph", {
-    ...hrTimesheetGraphView,
-    Renderer: ProjectTaskAnalysisGraphRenderer,
-});
