+++ addons/hr_timesheet/static/src/views/project_task_graph/project_task_graph_model.js
@@ -1,0 +1,4 @@
+import { ProjectTaskGraphModel } from "@project/views/project_task_graph/project_task_graph_model";
+import { patchGraphModel } from "../graph_model_patch";
+
+patchGraphModel(ProjectTaskGraphModel);
